requirements.txt
streamlit
transformers
langchain-openai
langchain-community
langchain-huggingface
python-dotenv
pillow
torch